// packages/ballerina-extension/src/utils/ai/types.ts
/**
 * Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com) All Rights Reserved.
 *
 * WSO2 LLC. licenses this file to you under the Apache License,
 * Version 2.0 (the "License"); you may not use this file except
 * in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

export interface CompletionRequest {
  prefix: string;
  suffix: string;
  maxCompletions?: number;
}

export interface CompletionResponse {
  completion: string;
}

export interface GithubCompletionRequest {
  prompt: string;
  suffix: string;
  n: number;
  max_tokens: number;
  temperature: number;
  top_p: number;
  stop: string[];
  nwo: string;
  stream: boolean;
  extra: {
    language: string;
    next_indent: number;
    trim_by_indentation: boolean;
  };
}

export interface CompletionChoice {
  text: string;
  index: number;
}

export interface CompletionChunk {
  choices: CompletionChoice[];
}